Raw Code
Raw code for Sankey Diagram. Just the .R file, no Jupyter Notebook. Also contains a couple alternative diagrams compared to the notebook, such as changing path color to all grey (and keeping node colors) and changing path color to mirror that of the source nodes instead of the target nodes.